+++ nope/lib/models/Content.php
@@ -23,17 +23,16 @@ class Content extends Model {
   function jsonSerialize() {
     $json = parent::jsonSerialize();
     $author = $this->getAuthor();
-    unset($json->author_id);
+    unset($json->authorId);
     $json->author = $author;
-    $json->id = (int) $json->id;
     $json->tags = $this->getTags();
-    if($json->cover_id) {
-      $cover = Media::findById($json->cover_id);
+    if($json->coverId) {
+      $cover = Media::findById($json->coverId);
       if($cover) {
         unset($cover->model->author_id);
       }
     }
-    unset($json->cover_id);
+    unset($json->coverId);
     $json->cover = $cover;
     $a = (object) $json;
     unset($a->sharedTag);
@@ -62,14 +61,13 @@ class Content extends Model {
 
   function setAuthor($user) {
     if($user->id) {
-      $this->model->author = R::load(User::MODELTYPE,$user->id);
+      $this->model->author = R::load(User::__getModelType(),$user->id);
     }
   }
 
   private function getAuthor() {
-    if($this->model->author_id) {
-      return User::findByid($this->model->author_id);
-      return $this->model->fetchAs(User::MODELTYPE)->author;
+    if($this->author_id) {
+      return User::findByid($this->author_id);
     }
     return null;
   }
@@ -111,11 +109,15 @@ class Content extends Model {
           $this->setCover($body[$f]);
         } else {
           if(array_key_exists($f, $body)) {
-            $this->model->$f = $body[$f];
+            $this->$f = $body[$f];
           }
         }
       }
     }
+  }
+
+  static public function findBySlug($slug) {
+    return self::__to(R::findOne(self::__getModelType(), 'slug = ?', [$slug]));
   }
 
 }