+++ nope/lib/models/User.php
@@ -28,8 +28,8 @@ class User extends \Nope\Model {
     $obj = parent::jsonSerialize();
     $obj->enabled = (int) $obj->enabled;
     $obj->permissions = $this->getPermissions();
-    if($obj->cover_id) {
-      $cover = Media::findById($obj->cover_id);
+    if($obj->coverId) {
+      $cover = Media::findById($obj->coverId);
       if($cover) {
         unset($cover->model->author_id);
       }
@@ -37,9 +37,9 @@ class User extends \Nope\Model {
     $obj->cover = $cover;
     unset($obj->password);
     unset($obj->salt);
-    unset($obj->last_login_date);
-    unset($obj->reset_code);
-    unset($obj->cover_id);
+    unset($obj->lastLoginDate);
+    unset($obj->resetCode);
+    unset($obj->coverId);
     return $obj;
   }
 
@@ -155,7 +155,7 @@ class User extends \Nope\Model {
     if(!$this->id && $userCheck) {
       $e = new \Exception("Error saving user due to existing username.");
       throw $e;
-    } elseif(!$this->id && $emailCheck && $this->email != '') {
+    } elseif(!$this->id && $emailCheck) {
       $e = new \Exception("Error saving user due to existing email.");
       throw $e;
     }