+++ nope/lib/routes/service.php
@@ -6,7 +6,7 @@ use Intervention\Image\ImageManagerStatic as Image;
 
 $app->group(NOPE_ADMIN_ROUTE . '/service', function() {
 
-  $this->get('/preview/{type:[a-z]+}/{id:[0-9]+}', function($req, $res, $args) {
+  $this->get('/preview/{type:[a-z]+}/{id:[0-9]+}', function($request, $response, $args) {
     $media = Media::findById($args['id']);
     $filter = \Nope::getConfig('nope.media.size')[$args['type']];
     if($media->isImage()) {
@@ -23,9 +23,9 @@ $app->group(NOPE_ADMIN_ROUTE . '/service', function() {
     $img = Image::cache(function($image) use ($path, $filter) {
       $image->make($path)->filter($filter);
     });
-    $body = $res->getBody();
+    $body = $response->getBody();
     $body->write($img);
-    return $res->withHeader('Content-Type', 'image/jpeg')
+    return $response->withHeader('Content-Type', 'image/jpeg')
       ->withHeader('Content-Length', strlen($img))
       ->withBody($body);
   });